tarfs: Normalize paths
In New, and then again in Open before we attempt to lookup the path in
the files map.  This allows us to use tarballs generated with GNU's
tar, which prefixes files with ./:

  $ tar --version | head -n1
  tar (GNU tar) 1.29
  $ mkdir a
  $ cd a
  $ touch readme.txt
  $ tar -cvf ../test.tar .
  ./
  ./readme.txt
  $ tar -tf ../test.tar
  ./
  ./readme.txt

When using that tarball for file URIs, we want file:///readme.txt to
resolve to that file, regardless of whether it's in the tarball as
readme.txt, /readme.txt, ./readme.txt, etc.